refactor: Phase 4.3 — shared types workspace
Introduce @flaque/shared as a pure-types workspace package holding
Track, TrackTags, TrackTagExtraValue, Playlist, PlaylistVisibility,
and LibraryIndex. Backend and frontend now re-export from this single
source of truth instead of maintaining duplicate, drifting definitions.

Playlist gains an optional trackCount?: number (previously frontend-only)
so both sides share the same shape.
